requirements.txt
torch~=2.4.0
flask~=3.0.3
googletrans~=3.0.0
pydub~=0.25.1
streamlit~=1.37.1
setuptools~=72.1.0
-e .
mysql~=0.0.3
mysql-connector-python~=9.0.0
pygame~=2.6.0
python-dotenv~=1.0.1
playsound~=1.3.0
whisper